Divides the app to moldules.
